Fix empty Project-Hash issue
The Project-Hash is allowed to be empty.
When calling IterationCollection.get_tests_overview, the empty string
was preserved, but when saving the PassedFailed CSV in between, it used
to be converted to None upon re-reading it, causing projects with empty
Project-Hashes to be dropped by the groupby. This commit fixes this
issue.